feat: Generate QR codes using app URL
This commit modifies the QR code generation process to use the
application's URL instead of just the item unit UUID. This ensures
that the generated QR codes will redirect to the item unit's page
within the application.  Also forces http for local development.

Updates the GenerateItemUnitQRsBatch job to construct a URL; Refs: #123